pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mova"
version = "0.1.0"
description = "MOVA codebase"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "MOVA contributors" }]

dependencies = [
  "torch",
  "torchvision",
  "pillow",
  "tqdm",
  "imageio",
  "imageio-ffmpeg",
  "accelerate",
  "diffusers",
  "transformers<5",
  "mmengine",
  "yunchang",
  "ftfy",
  "descript-audiotools",
  "einops",
]

[project.optional-dependencies]
# Training-only dependencies. Install via: `pip install .[train]`
train = [
  "bitsandbytes",
  "torchcodec",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["mova*"]

